Split the synaptic manager into application and machine concerns
