+++ spynnaker/pyNN/models/neuron/__init__.py
@@ -16,11 +16,10 @@
 from .abstract_population_vertex import AbstractPopulationVertex
 from .connection_holder import ConnectionHolder
 from .population_machine_vertex import PopulationMachineVertex
-from .synaptic_manager import SynapticManager
 from .abstract_pynn_neuron_model import AbstractPyNNNeuronModel
 from .abstract_pynn_neuron_model_standard import (
     AbstractPyNNNeuronModelStandard)
 
 __all__ = ["AbstractPopulationVertex", "AbstractPyNNNeuronModel",
            "AbstractPyNNNeuronModelStandard", "ConnectionHolder",
-           "PopulationMachineVertex", "SynapticManager"]
+           "PopulationMachineVertex"]
