+++ unittests/model_tests/neuron/test_synapse_io.py
@@ -19,8 +19,6 @@ from spynnaker.pyNN.models.neural_projections import (
     ProjectionApplicationEdge, SynapseInformation)
 from spynnaker.pyNN.models.neuron.synapse_dynamics import (
     SynapseDynamicsStatic, SynapseDynamicsSTDP)
-from spynnaker.pyNN.models.neuron.master_pop_table import (
-    MasterPopTableAsBinarySearch)
 from spynnaker.pyNN.models.neuron.synapse_io import _get_allowed_row_length
 from spynnaker.pyNN.models.neuron.plasticity.stdp.weight_dependence import (
     WeightDependenceAdditive)
@@ -55,16 +53,13 @@ def test_get_allowed_row_length(
         dynamics = dynamics_class(timing(), weight())
     else:
         dynamics = dynamics_class()
-    population_table = MasterPopTableAsBinarySearch()
     synapse_information = SynapseInformation(
         None, None, None, False, False, None, None, dynamics, 0, True)
     in_edge = ProjectionApplicationEdge(None, None, synapse_information)
     if exception is not None:
         with pytest.raises(exception) as exc_info:
-            _get_allowed_row_length(
-                size, dynamics, population_table, in_edge, size)
+            _get_allowed_row_length(size, dynamics, in_edge, size)
         assert exc_info.value.max_size == max_size
     else:
-        actual_size = _get_allowed_row_length(
-            size, dynamics, population_table, in_edge, size)
+        actual_size = _get_allowed_row_length(size, dynamics, in_edge, size)
         assert actual_size == max_size
